+++ cli/src/commands/parse.rs
@@ -8,8 +8,11 @@ use std::{env, fs, io};
 
 use anyhow::{anyhow, Context, Result};
 use phylum_lockfile::generator::Generator;
-use phylum_lockfile::{LockfileFormat, Package, PackageVersion, Parse, ThirdPartyVersion};
+use phylum_lockfile::{
+    LockfileFormat, Package as LockfilePackage, PackageVersion, Parse, ThirdPartyVersion,
+};
 use phylum_types::types::package::{PackageDescriptor, PackageDescriptorAndLockfile};
+use serde::{Deserialize, Serialize};
 
 use crate::commands::{CommandResult, ExitCode};
 use crate::{config, print_user_failure, print_user_warning};
@@ -24,23 +27,28 @@ pub enum ParseError {
     Other(#[from] anyhow::Error),
 }
 
+#[derive(Serialize, Deserialize, Debug)]
 pub struct ParsedLockfile {
-    pub packages: Vec<PackageDescriptorAndLockfile>,
+    pub packages: Vec<PackageDescriptor>,
     pub format: LockfileFormat,
     pub path: PathBuf,
 }
 
 impl ParsedLockfile {
     fn new(path: PathBuf, format: LockfileFormat, packages: Vec<PackageDescriptor>) -> Self {
-        let packages = packages
-            .into_iter()
-            .map(|package_descriptor| PackageDescriptorAndLockfile {
-                package_descriptor,
-                lockfile: Some(path.to_string_lossy().into_owned()),
-            })
-            .collect();
-
         Self { packages, format, path }
+    }
+
+    /// Convert packages to API's expected format.
+    pub fn api_packages(&self) -> Vec<PackageDescriptorAndLockfile> {
+        let lockfile = Some(self.path.to_string_lossy().into_owned());
+        self.packages
+            .iter()
+            .map(|package_descriptor| PackageDescriptorAndLockfile {
+                package_descriptor: package_descriptor.clone(),
+                lockfile: lockfile.clone(),
+            })
+            .collect()
     }
 }
 
@@ -61,14 +69,14 @@ pub fn handle_parse(matches: &clap::ArgMatches) -> CommandResult {
 
     let project = phylum_project::get_current_project();
     let project_root = project.as_ref().map(|p| p.root());
-    let lockfiles = config::lockfiles(matches, project.as_ref())?;
+    let depfiles = config::depfiles(matches, project.as_ref())?;
 
-    let mut pkgs: Vec<PackageDescriptorAndLockfile> = Vec::new();
-    for lockfile in lockfiles {
-        let parse_result = parse_lockfile(
-            &lockfile.path,
+    let mut pkgs = Vec::new();
+    for depfile in depfiles {
+        let parse_result = parse_depfile(
+            &depfile.path,
             project_root,
-            Some(&lockfile.lockfile_type),
+            Some(&depfile.depfile_type),
             sandbox_generation,
             generate_lockfiles,
         );
@@ -77,12 +85,12 @@ pub fn handle_parse(matches: &clap::ArgMatches) -> CommandResult {
         let mut parsed_lockfile = match parse_result {
             Ok(parsed_lockfile) => parsed_lockfile,
             Err(err @ ParseError::ManifestWithoutGeneration(_)) => {
-                print_user_failure!("Could not parse lockfile: {}", err);
+                print_user_failure!("Could not parse manifest: {}", err);
                 return Ok(ExitCode::ManifestWithoutGeneration);
             },
             Err(ParseError::Other(err)) => {
                 return Err(err).with_context(|| {
-                    format!("could not parse lockfile: {}", lockfile.path.display())
+                    format!("could not parse dependency file \"{}\"", depfile.path.display())
                 });
             },
         };
@@ -95,17 +103,17 @@ pub fn handle_parse(matches: &clap::ArgMatches) -> CommandResult {
     Ok(ExitCode::Ok)
 }
 
-/// Parse a package lockfile.
-pub fn parse_lockfile(
+/// Parse a dependency file.
+pub fn parse_depfile(
     path: impl Into<PathBuf>,
     project_root: Option<&PathBuf>,
-    lockfile_type: Option<&str>,
+    depfile_type: Option<&str>,
     sandbox_generation: bool,
     generate_lockfiles: bool,
 ) -> StdResult<ParsedLockfile, ParseError> {
-    // Try and determine lockfile format.
+    // Try and determine dependency file format.
     let path = path.into();
-    let format = find_lockfile_format(&path, lockfile_type);
+    let format = find_depfile_format(&path, depfile_type);
 
     // Attempt to parse with all known parsers as fallback.
     let (format, lockfile) = match format {
@@ -146,7 +154,7 @@ pub fn parse_lockfile(
 
     // If the lockfile couldn't be parsed, or there is none, we generate a new one.
 
-    // Find the generator for this lockfile format.
+    // Find the generator for this format.
     let generator = match parser.generator() {
         Some(generator) => generator,
         None => return Err(anyhow!("unsupported manifest file {path:?}").into()),
@@ -209,14 +217,14 @@ fn parse_lockfile_content(
     Ok(filter_packages(packages))
 }
 
-/// Find a lockfile's format.
-fn find_lockfile_format(
+/// Find a dependency file's format.
+fn find_depfile_format(
     path: &Path,
-    lockfile_type: Option<&str>,
+    depfile_type: Option<&str>,
 ) -> Option<(LockfileFormat, Option<PathBuf>)> {
-    // Determine format from lockfile type.
-    if let Some(lockfile_type) = lockfile_type.filter(|lockfile_type| lockfile_type != &"auto") {
-        let format = lockfile_type.parse::<LockfileFormat>().unwrap();
+    // Determine format from dependency file type.
+    if let Some(depfile_type) = depfile_type.filter(|depfile_type| depfile_type != &"auto") {
+        let format = depfile_type.parse::<LockfileFormat>().unwrap();
 
         // Skip lockfile analysis when path is only valid manifest.
         let parser = format.parser();
@@ -226,7 +234,7 @@ fn find_lockfile_format(
         return Some((format, lockfile));
     }
 
-    // Determine format based on lockfile path.
+    // Determine format based on dependency file path.
     if let Some(format) = phylum_lockfile::get_path_format(path) {
         return Some((format, Some(path.into())));
     }
@@ -286,7 +294,7 @@ fn try_get_packages(path: PathBuf, project_root: Option<&PathBuf>) -> Result<Par
 }
 
 /// Filter packages for submission.
-fn filter_packages(mut packages: Vec<Package>) -> Vec<PackageDescriptor> {
+fn filter_packages(mut packages: Vec<LockfilePackage>) -> Vec<PackageDescriptor> {
     packages
         .drain(..)
         .filter_map(|package| {